services/post-incident-agent/pyproject.toml
[project]
name = "post-incident-agent"
version = "0.1.0"
description = "A post-incident analysis agent that learns from past incidents, evaluates response effectiveness, identifies patterns, and provides system optimization recommendations using BigQuery, Vertex AI, and Cloud Analytics."
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.8.0,<2.0.0",
    "google-cloud-aiplatform[adk,agent_engine]>=1.100.0",
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-storage>=2.0.0",
    "google-cloud-logging>=3.0.0",
    "python-dotenv>=1.0.0",
    "pathlib",
    "asyncio",
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    "dash>=2.0.0",
    "sqlalchemy>=1.4.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false 